+++ ECommerceApi/MODEL/KangJiaQi/BUser/RoleInfo.cs
@@ -14,7 +14,7 @@ namespace MODEL
         /// <summary>
         /// 角色id
         /// </summary>
-        public int  RoleId     { get; set; }
+        public long RoleId     { get; set; }
         /// <summary>
         /// 角色名称
         /// </summary>
@@ -30,10 +30,10 @@ namespace MODEL
         /// <summary>
         /// 创建人
         /// </summary>
-        public int CreateId { get; set; }
+        public long CreateId { get; set; }
         /// <summary>
         /// 修改人
         /// </summary>
-        public int UpdateId { get; set; }
+        public long UpdateId { get; set; }
     }
 }
